requirements_sub.txt
flask
mysql-connector-python
requests
flasgger
python-dotenv
flask_sqlalchemy
pymysql
